feat: Clean up post request function
